+++ app.js
@@ -1,15 +1,11 @@
-var createError = require('http-errors');
-var express = require('express');
-var path = require('path');
-var cookieParser = require('cookie-parser');
-var logger = require('morgan');
+const createError = require('http-errors');
+const express = require('express');
+const path = require('path');
+const cookieParser = require('cookie-parser');
+const logger = require('morgan');
 const cors = require('cors');
-const errorController = require('./controllers/errorController');
-
-// routers
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
-const authRouter = require('./routes/auth');
+const errorMiddleware = require('./middlewares/errorMiddleware');
+const rootRouter = require('./routes/rootRouter');
 
 // init db
 const sequelize = require('./models/db');
@@ -20,15 +16,11 @@ const models = require('./models/models');
     await sequelize.sync();
     console.log('DB connection has been established successfully.');
   } catch (err) {
-    console.error('Unable to connect to the database:', error);
+    console.log('Unable to connect to the database:', err);
   };
 })();
 
-var app = express();
-
-// view engine setup
-// app.set('views', path.join(__dirname, 'views'));
-// app.set('view engine', 'jade');
+const app = express();
 
 app.use(cors());
 app.use(logger('dev'));
@@ -37,15 +29,8 @@ app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
 app.use(express.static(path.join(__dirname, 'public')));
 
-app.use('/', indexRouter);
-app.use('/users', usersRouter);
-app.use('/auth', authRouter);
+app.use('/api', rootRouter);
 
-// catch 404 and forward to error handler
-// app.use(function(req, res, next) {
-//   next(createError(404));
-// });
-// error handler
-app.use(errorController.handleError);
+app.use(errorMiddleware.handleError);
 
 module.exports = app;
