--- routes/index.js
@@ -1,10 +1,0 @@
-var express = require('express');
-var router = express.Router();
-
-/* GET home page. */
-router.get('/', function(req, res, next) {
-  console.log('get home page');
-  res.render('index', { title: 'Express' });
-});
-
-module.exports = router;
